requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# Image processing
scikit-image>=0.21.0
opencv-python-headless>=4.8.0
imagecodecs>=2023.1.0

# Data handling
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Scientific computing
scipy>=1.10.0

# Optional: For augmentation
# albumentations>=1.3.0
